pyproject.toml
[project]
name = "gol-reportnik-prau-271"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.1.1",
]
